+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.12] - 2020-03-08
+## [0.0.13] - 2020-03-08
 ### Added
 * Typescript definition files